+++ index.html
@@ -43,5 +43,6 @@
     }
   </style>
 
- <script src="index.js"></script>
+ <script type="module" src="main.js"></script>
+ 
 </html>
